+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- hnefatafl-server-full: archive games to file.
+- hnefatafl-server-full: archive games to file and load from file.
 
 ## 1.2.3 - 2025-06-07